Refactor: 아이디, 비밀번호 찾기 타입을 하나로 통합
