--- src/types/changePasswordFormTypes.ts
@@ -1,4 +1,0 @@
-export interface ChangePasswordFormProps {
-  password: string;
-  password_confirm: string;
-}
