--- src/types/passwordTypes.ts
@@ -1,8 +1,0 @@
-export interface FindPasswordInputs {
-  [index: string]: string;
-}
-export interface RecoveryUser {
-  name: string;
-  email: string;
-  phoneNumber: string;
-}
